refactor: status 관련 텍스트들 한파일로 합침
