+++ apps/web/app/books/[id]/_components/ReadingCard.tsx
@@ -4,9 +4,9 @@ import { Chip } from '@repo/ui/components/Chip';
 import { Icon } from '@repo/ui/components/Icon';
 import { Box, HStack, JustifyBetween, Stack } from '@repo/ui/components/Layout';
 import { Text } from '@repo/ui/components/Text';
+import { MY_READING_STATUS_TEXT } from 'app/_constants/status';
 
 import { MOCK_TAG } from '../_fixture/hashTag';
-import { STATUS_TEXT } from '../status';
 
 import { DialogTrigger } from './DialogTrigger';
 
@@ -15,10 +15,15 @@ export const ReadingCard = () => {
     <Stack className="gap-5 rounded-lg bg-white px-5 py-4">
       <JustifyBetween className="items-center">
         <HStack className="gap-3">
-          <Image src={STATUS_TEXT['prev'].image} alt="독서 상태" width={52} height={52} />
+          <Image
+            src={MY_READING_STATUS_TEXT['prev'].image}
+            alt="독서 상태"
+            width={52}
+            height={52}
+          />
           <Stack className="gap-0.5">
             <Text type="Title1" weight="semibold" className="text-gray-900">
-              {STATUS_TEXT['prev'].text}
+              {MY_READING_STATUS_TEXT['prev'].text}
             </Text>
             <Text type="Body2" weight="medium" className="text-gray-400">
               2025.01.14
